Definitely removed all references to Tao* stuff from sipxcalllib. Tao code was removed from SVN.
ISocketEvent was removed from CpMediaInterface as this is not how we should signal events.

addHistoryEvent was removed, as its content wasn't written anywhere, only stored in CpCall, and this will be redone.

Mess in includes in several .h files was cleaned up.

CpMultiStringMessage.h was reformatted.

CpNotificationMsgDef.h was created to contain definitions that are required in several files, but keeping it elsewhere would introduce bad dependencies.

Still more cleanup is required before sipxcalllib can be refactored.