Merged revisions 9987,9988 from SIP foundry:
-------------------------------------------

Fix stricmp C4996 warnings on VS2005 -- use ISO C++ conformant name instead of POSIX name as it is deprecated.

fixes, among others:
SpeakerThreadWnt.cpp
.\src\mp\SpeakerThreadWnt.cpp(367) : warning C4996: 'stricmp' was declared deprecated
        C:\Program Files\Microsoft Visual Studio 8\VC\include\string.h(213) : see declaration of 'stricmp'
        Message: 'The POSIX name for this item is deprecated. Instead, use the ISO C++ conformant name: _stricmp. See online help for details.'

-------------------------------------------------------

OsDefs stricmp change for vs2005 should be checking for _MSC_VER >= 1400 instead of _MSC_VER > 1300, as MS may increase 1300 to something between 1300 and 1400 for a service pack to vs2003.  So, more correct is to check against 1400.