Improvements in sipxLineAdd sipxtapi method. Actually supplied contactId is used only to select contactType and not IP or transport. Since actual contactType used can be overridden internally, save only selected contactType, address, port and transport into line instead of contactId. ContactId in line is a useless information.
API Change:
- removed function sipxLineGetContactId as it provided useless data
- added function sipxLineGetContactInfo which can provide contact address, port, contactType.

When using TCP or TLS, contact may or may not use TCP or TLS port depending on contactType of supplied contactId. When using local contact, correct TCP or TLS port will be used by contact, but a local IP address will be used. When using automatic or nat mapped contact, then a UDP port will be used by the contact. If UDP port is the same as TCP port (by default it is) then the port will be correct.