Fixed a couple of bugs introduced in SipDialog. Bugs were fixed in isSameDialog and isInitialDialogFor. Dialog matching was too string and incorrect. Now it is exactly as it should be, not as lenient as it was in original implementation, where there was confusion between local/remote from/to tags and they needed to be swapped if one solution didn't work. Now we know exactly if they have to be swapped or not.
Moved getDialogHandle from SipMessage into SipDialog. SipMessage doesn't need to know anything about a dialog. Dialog state is now updated better. SipMessage cannot construct dialog handles, because it doesn't know if tags need to be swapped. Swapping tags is needed so that dialog handle uniquely identifies a dialog, regardless if its inbound/outbound request/response. Previously it was not the case, dialog handle was not unique and it brought mess into the code. Code which relied on the fact that from tag was always first will be obviously broken.