Fixed sipxtacklib unit tests. Now all unit tests pass.
Added new unit tests for SipDialog to keep it stable and test new functionality.
Some tests had to be fixed due to change of how Url.toString() works, now it will never print <> for request uri, and print <> for normal uri by default, unless disabled. This broke a couple of tests.
In some tests SipMessage needs to be marked with flag isFromThisSide correctly, for SipDialog to work.