Attempt to get rid of CallManager::setOutboundLineForCall. Now CpPeerCall doesn't contain local sip uri as before. CallManager has no default local sip uri.
local sip uri is stored in SipConnection, so that it can be reused if needed.

SipLineProvider now automatically detects if SipMessage is inbound or outbound according to isFromThisSide SipMessage flag.

Added parameters into Url::getUri to optionally remove display name or brackets. This method probably originally constructed only the sip request uri, but then due to confusion founds its way into all code.

SipLine is now capable of constructing lineURI from Url or UtlString, in order to unify the way it is constructed all over the code.

There were quite big changes in sipXcallLib and as a result not everything is working. Only basic call setup from local/remote side and teardown works, but INVITE tags are not generated correctly. But this was necessary to clean up the confusion in sipXcallLib regarding used fromUrls.