Cleaned up URL/URI confusion in many parts of the code. SIP URI may not contain display name, field parameters or <> SIP URL is made of optional display name, optional field parameters, and optional <> around SIP URI
Renamed a couple of fields to adhere to this rule.

In SipLine, we now use fullLineUrl, which is almost the original url, except that we strip field parameters, and remove "transport" parameter (transport will be set differently and later). The reason for using different way of transport selection is support for custom transport.

Fields in SIPX_CALL_DATA have been renamed to begin with m_.