Added support for choice if we want to send SDP in 18x responses when accepting call. This choice is respected only if unreliable 18x is sent, or reliable 18x is sent and initial INVITE contained SDP offer. If INVITE didn't contain SDP offer and we are sending reliable 18x, we must add an SDP offer.
Added support for 100rel for inbound calls. We generate reliable 18x responses, with SDP if necessary, and process PRACKs with SDP (offer or answer) if necessary. If we detect some error during SDP negotiation when processing reliable 18x response, we try to complete the call and then terminate it with BYE with Reason.