Changed log level of even handlers to debug. There is no need to log so much data in INFO level.
Removed code checking for duplicate audio events. There are no duplicate audio events sent from sipxcalllib now, and this code is not needed. The only weird events that can be seen is that it is possible to get MEDIA_LOCAL_START, MEDIA_REMOTE_START, and then a little bit later, the same 2 events again, without seeing STOP. That happens when there is re-INVITE or UPDATE and decoder/encoder is reset with new codecs. This also allows to detect a re-INVITE.