First functional version of sipxConferenceJoin and sipxConferenceSplit. After sipxConferenceSplit, split call remains in bridged state (because conference retains focus).
When XSipConnection moves between call and conference, it might happen that some timer or sip messages are left behind in the wrong queue. But they will be reposted into call manager who will then route them to the correct call/conference.

Timers in SipConnectionStateContext are not yet restarted, so they might send a message into deleted queue. This will be fixed later.