Fixed bug with PRACK not being generated if INVITE contained SDP but reliable 18x did not. SDP doesn't have to be in reliable 18x if INVITE contained SDP according to RFC 3262.
Fixed problem with sipXcallAccept bSendSdp being ignored when INVITE contained SDP and 100rel was enabled. In that case we may choose to send or not send SDP in reliable 18x.

Fixed problem with 18x not being processed for re-INVITEs. This would result in PRACKs not being sent for reliable 183 and SDP answers being ignored.